Uploading PNGS sequence sets
Sequences 'check', 'idle', 'bpm', and 'noise' are labeled with the sequence type, followed by the 'stage' #, as well as 'A'/'B' if the PNG reflects the result of an input-dependant choice. If the PNG is the last PNG to appear in a pathway, indicating the end of the sequence, the 'stage' # will be followed by '_END'.  

Sequence 'wake' is labeled with the old naming system.